+++ SharpCoreSource/SharpCore/src/Ticking/Utility/TickExecutorUtility.cs
@@ -13,11 +13,11 @@ namespace SharpCore.Ticking
         /// </summary>
         /// <param name="delta">Delta since last tick (seconds).</param>
         /// <param name="simulationTicks">The simulation ticks on which to execute.</param>
-        public static void ExecuteSimulationTicks(double delta, IEnumerable<TickSimulation> simulationTicks)
+        public static void ExecuteSimulationTicks(float delta, IEnumerable<TickSimulation> simulationTicks)
         {
             foreach (TickSimulation sim in simulationTicks)
             {
-                double tickrate = sim.configData.tickrate;
+                float tickrate = sim.configData.tickrate;
                 sim.accumulator += delta;
                 sim.accumulator = 
                     Math.Min(sim.accumulator, sim.configData.maxDelta);
@@ -35,7 +35,7 @@ namespace SharpCore.Ticking
         /// </summary>
         /// <param name="delta">Delta since last tick (seconds).</param>
         /// <param name="renderTick">The render tick on which to operate.</param>
-        public static void ExecuteRenderTicks(double delta, ITickInstance<ITickRenderClient> renderTick)
+        public static void ExecuteRenderTicks(float delta, ITickInstance<ITickRenderClient> renderTick)
         {
             renderTick.Tick(delta);
         }
